code cleanup
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@15220 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/javascript/host/dom/Document.java
@@ -1285,7 +1285,7 @@ public class Document extends Node {
             boolean expandEntityReferences) throws DOMException {
 
         // seems that Rhino doesn't like long as parameter type
-        final int whatToShowI = (int) Double.valueOf(whatToShow).longValue();
+        final int whatToShowI = (int) whatToShow;
 
         if (getBrowserVersion().hasFeature(JS_TREEWALKER_EXPAND_ENTITY_REFERENCES_FALSE)) {
             expandEntityReferences = false;
